# shared/openapi/openapi.yaml
openapi: "3.1.0"
info:
  title: EDO API Gateway
  version: "0.1.0"
  description: |
    BFF (Backend For Frontend) Gateway for the Electronic Document Management System.
    Handles session management, Keycloak auth, and proxies requests to microservices via gRPC.
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
  /api/documents:
    post:
      tags: [documents]
      summary: Create document draft
      operationId: createDocumentDraft
      description: Routed via gateway to document-service during migration wave 1.
      x-migration-wave: wave-1-document-service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentRequest"
      responses:
        "201":
          description: Draft document created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentResponse"
    get:
      tags: [documents]
      summary: Search documents
      operationId: searchDocuments
      description: Backward-compatible route; provider transitions from draft service to search-notification-service.
      x-migration-wave: wave-4-search-notification-service
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: scope
          schema:
            type: string
            enum: [mine]
          description: Restrict results to documents where the current user is owner/author or workflow approver.
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/DocumentResponse"
...
  /api/tasks/{taskId}:
    get:
      tags: [documents]
      summary: Get task details
      operationId: getTaskDetails
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
        - in: query
          name: boardId
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Task details with board context and capabilities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDetailsResponse"
...
  /api/boards:
    post:
      tags: [documents]
      summary: Create a task board
      operationId: createTaskBoard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBoardRequest"
      responses:
        "201":
          description: Task board created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTaskBoardResponse"
        "400":
          description: Invalid board request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      tags: [documents]
      summary: List task boards for an organization
      operationId: listTaskBoards
      parameters:
        - in: query
          name: organizationId
          required: true
          schema:
            type: string
          description: Organization identifier
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
            maximum: 100
          description: Maximum number of boards to return
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          description: Number of boards to skip
      responses:
        "200":
          description: List of task boards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskBoardsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/boards/{boardId}:
    get:
      tags: [documents]
      summary: Get detailed information about a task board
      operationId: getTaskBoard
      parameters:
        - in: path
          name: boardId
          required: true
          schema:
            type: string
          description: Board identifier
      responses:
        "200":
          description: Task board details with tasks and available resources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskBoardDetailsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Board not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/activity:
    get:
      tags: [documents]
      summary: Get system activity history
      operationId: listActivityEvents
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Activity history
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [id, actorUserId, actorUserName, entityType, entityId, actionType, summary, occurredAt]
                      properties:
                        id:
                          type: string
                        actorUserId:
                          type: string
                        actorUserName:
                          type: string
                        entityType:
                          type: string
                          enum: [DOCUMENT, TASK]
                        entityId:
                          type: string
                        actionType:
                          type: string
                        summary:
                          type: string
                        occurredAt:
                          type: string
                        documentId:
                          type: string
                        taskId:
                          type: string
                        boardId:
                          type: string
                  total:
                    type: integer
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "Unauthorized"
    CreateDocumentRequest:
      type: object
      required: [title, category]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 300
        category:
          type: string
          enum: [HR, FINANCE, GENERAL]
...
    DocumentStatus:
      type: string
      enum: [DRAFT, IN_REVIEW, CHANGES_REQUESTED, APPROVED, ARCHIVED]
    DocumentCapabilities:
      type: object
      required: [canEdit, canSubmit, canApprove, canRequestChanges, canArchive]
      properties:
        canEdit:
          type: boolean
        canSubmit:
          type: boolean
        canApprove:
          type: boolean
        canRequestChanges:
          type: boolean
        canArchive:
          type: boolean
...
    DocumentResponse:
      allOf:
        - $ref: "#/components/schemas/DocumentCapabilities"
        - type: object
          required: [id, title, category, status, version, updatedAt]
          properties:
            id:
              type: string
              format: uuid
            title:
              type: string
            category:
              type: string
            status:
              $ref: "#/components/schemas/DocumentStatus"
            contentDocument:
              type: object
              additionalProperties: true
            version:
              type: integer
              minimum: 1
            updatedAt:
              type: string
              format: date-time
            objectKey:
              type: string
            objectVersionId:
              type: string
...
    TaskType:
      type: string
      enum: [approval, general]
      description: Type of task - general tasks or those requiring approval
    TaskStatus:
      type: string
      enum: [pending, in_review, approved, declined]
      description: Current status of the task in the workflow
    TaskDecision:
      type: string
      enum: [approved, declined]
      description: Decision made by the approver
    KanbanTaskGroupBy:
      type: string
      enum: [assignee, department, group]
      description: How to group tasks in the Kanban view
...
    TaskAttachment:
      type: object
      required: [documentId, title, category, status]
      properties:
        documentId:
          type: string
          description: Document identifier
        title:
          type: string
          description: Document title
        category:
          type: string
          enum: [HR, FINANCE, GENERAL]
          description: Document category
        status:
          $ref: "#/components/schemas/DocumentStatus"
          description: Document status
    TaskComment:
      type: object
      required: [id, authorId, authorName, text, createdAtLabel]
      properties:
        id:
          type: string
          description: Comment identifier
        authorId:
          type: string
          description: Author user ID
        authorName:
          type: string
          description: Author display name
        text:
          type: string
          description: Comment text
        createdAtLabel:
          type: string
          description: Human-readable creation timestamp
    KanbanTask:
      type: object
      required: [id, title, status, assigneeId, assigneeName, department, groupId, groupName, dueDateLabel, comments, creatorId, creatorName, attachments, taskType, createdAt, updatedAt]
      properties:
        id:
          type: string
          description: Task identifier
        boardId:
          type: string
          nullable: true
          description: Board identifier for task details navigation
        title:
          type: string
          description: Task title
        description:
          type: string
          description: Task description
        status:
          $ref: "#/components/schemas/TaskStatus"
        assigneeId:
          type: string
          nullable: true
          description: Assigned user ID
        assigneeName:
          type: string
          description: Assigned user display name
        department:
          type: string
          description: User's department
        groupId:
          type: string
          description: Task group identifier
        groupName:
          type: string
          description: Task group display name
        dueDateLabel:
          type: string
          description: Human-readable due date
        comments:
          type: array
          items:
            $ref: "#/components/schemas/TaskComment"
        creatorId:
          type: string
          description: User who created the task
        creatorName:
          type: string
          description: Creator display name
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/TaskAttachment"
        approverId:
          type: string
          nullable: true
          description: Approver user ID
        approverName:
          type: string
          nullable: true
          description: Approver display name
        taskType:
          $ref: "#/components/schemas/TaskType"
        decision:
          $ref: "#/components/schemas/TaskDecision"
          nullable: true
        decisionComment:
          type: string
          nullable: true
          description: Approver's decision comment
        createdAt:
          type: string
          format: date-time
          description: Task creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Task last update timestamp
...
    TaskDetailsResponse:
      type: object
      required: [board, task, members, currentUserId, canEdit, canAssign, canApprove, canMoveToReview, canComment]
      properties:
        board:
          $ref: "#/components/schemas/KanbanBoardSummary"
        task:
          $ref: "#/components/schemas/KanbanTask"
        members:
          type: array
          items:
            $ref: "#/components/schemas/KanbanBoardMember"
        currentUserId:
          type: string
        canEdit:
          type: boolean
        canAssign:
          type: boolean
        canApprove:
          type: boolean
        canMoveToReview:
          type: boolean
        canComment:
          type: boolean
    KanbanBoardMember:
      type: object
      required: [id, fullName, department]
      properties:
        id:
          type: string
          description: User identifier
        fullName:
          type: string
          description: User's full name
        department:
          type: string
          description: User's department
    KanbanBoardSummary:
      type: object
      required: [id, organizationId, name, description, membersCount, tasksCount]
      properties:
        id:
          type: string
          description: Board identifier
        organizationId:
          type: string
          description: Organization identifier
        name:
          type: string
          description: Board display name
        description:
          type: string
          description: Board description
        membersCount:
          type: integer
          minimum: 0
          description: Number of board members
        tasksCount:
          type: integer
          minimum: 0
          description: Number of tasks on the board
    TaskBoardsResponse:
      type: object
      required: [boards, total, page, pageSize]
      properties:
        boards:
          type: array
          items:
            $ref: "#/components/schemas/KanbanBoardSummary"
        total:
          type: integer
          minimum: 0
          description: Total number of boards
        page:
          type: integer
          minimum: 0
          description: Current page number
        pageSize:
          type: integer
          minimum: 1
          description: Number of boards per page
    CreateTaskBoardRequest:
      type: object
      required: [organizationId, name]
      properties:
        organizationId:
          type: string
          description: Organization identifier
        name:
          type: string
          minLength: 3
          maxLength: 200
          description: Board display name
        description:
          type: string
          description: Optional board description
    CreateTaskBoardResponse:
      type: object
      required: [board]
      properties:
        board:
          $ref: "#/components/schemas/KanbanBoardSummary"
    AvailableApproverItem:
      type: object
      required: [userId, userName]
      properties:
        userId:
          type: string
          description: User identifier
        userName:
          type: string
          description: User display name
...
    AvailableDocumentItem:
      type: object
      required: [documentId, title, category]
      properties:
        documentId:
          type: string
          description: Document identifier
        title:
          type: string
          description: Document title
        category:
          type: string
          enum: [HR, FINANCE, GENERAL]
          description: Document category
...
    TaskBoardDetailsResponse:
      type: object
      required: [id, organizationId, name, description, allowedGrouping, members, tasks, availableApprovers, availableDocuments]
      properties:
        id:
          type: string
          description: Board identifier
        organizationId:
          type: string
          description: Organization identifier
        name:
          type: string
          description: Board display name
        description:
          type: string
          description: Board description
        allowedGrouping:
          type: array
          items:
            $ref: "#/components/schemas/KanbanTaskGroupBy"
          description: Allowed grouping options for this board
        members:
          type: array
          items:
            $ref: "#/components/schemas/KanbanBoardMember"
          description: Board members
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/KanbanTask"
          description: Tasks on the board
        availableApprovers:
          type: array
          items:
            $ref: "#/components/schemas/AvailableApproverItem"
          description: Users who can be selected as approvers
        availableDocuments:
          type: array
          items:
            $ref: "#/components/schemas/AvailableDocumentItem"
          description: Documents available for attachment
...
  securitySchemes:
    keycloak:
      type: openIdConnect
      openIdConnectUrl: http://localhost:8080/realms/edo/.well-known/openid-configuration
    sessionCookie:
      type: apiKey
      in: cookie
      name: sessionId